requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
beautifulsoup4==4.12.2
duckdb==0.8.1
scikit-learn==1.3.0
gunicorn==21.2.0
lxml==4.9.3
html5lib==1.1
Pillow==10.0.1